Split material-group editor into create and edit
